代码审查修复 Wave 3 / P1-5：横向偏移走 64 位（ScrollMapper + qint64 extent）
问题：纵向早有 ScrollMapper，横向没有 ——
- syncHorizontalScrollBar() 把区间夹到 INT_MAX，然后把 QScrollBar 的 int 值写回
  HeaderGeometry::viewportOffset()，超过 INT_MAX 的横向偏移被静默截断；
- TablePaneLayout::extentOf() 与 ResolvedPane::extent 是 int，超宽表格会溢出；
- 离屏列的视口 x 被夹到 INT_MIN/INT_MAX，随后的 QRect/QWidget 运算（Qt 6.11 用 checked int）
  会直接断言崩溃；
- 原生 QHeaderView 自身内部就是 int 位置累加，内容范围超过 INT_MAX 时它自己会断言。

- 表格持有一个横向 ScrollMapper：滚动条只承载压缩值，逻辑偏移与回写都是 qint64，且窗口锚点跟随
  当前位置（与纵向一致）。
- pane 的 extent 全部 qint64（groupExtent / maximumGroupOffset 本来就是）。
- HeaderGeometry::columnGeometry() 与 TablePaneLayout 的 x 缓存改夹到"窗口附近哨兵范围"
  （±2^20）：离屏列仍然"不在窗口内"，但不会让 32 位 QRect 运算溢出。
- NativeHeaderView 在内容范围超出 int 时跳过镜像并 qWarning() 一次（注释说明这是 QHeaderView
  自身的限制），而不是让 Qt 断言。

回归测试 tst_tablepanes::horizontalOffsetSurvivesBeyondTheIntRange（30,000 列 x 100,000 px =
3e9 px）：逻辑内容范围与最大偏移都超出 INT_MAX；偏移能到最大、中点到中点精确；最后一列的右边缘
正好落在视口右边缘。

验证：Qt6 与 Qt5 两个配置全量 CTest（28 个目标）全绿。CHANGELOG 与 roadmap 同步。

+++ include/virtualitemviews/tablepane.h
@@ -200,7 +200,7 @@ public:
     QVector<int> columnsForLayout(int overscan) const;
 
 private:
-    int extentOf(const QVector<int> &logicalColumns) const;
+    qint64 extentOf(const QVector<int> &logicalColumns) const;
     QVector<int> visualOrderOf(const QVector<int> &logicalColumns) const;
     /// Applies \a remap (which may yield -1 for a dropped column) to the frozen
     /// sets and to every explicit pane spec, dropping what is gone.
